+++ bundles/org.eclipse.swt/Eclipse SWT PI/cocoa/org/eclipse/swt/internal/cocoa/NSColorPanel.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2000, 2012 IBM Corporation and others.
+ * Copyright (c) 2000, 2019 IBM Corporation and others.
  *
  * This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
@@ -19,7 +19,7 @@ public NSColorPanel() {
 	super();
 }
 
-public NSColorPanel(long /*int*/ id) {
+public NSColorPanel(long id) {
 	super(id);
 }
 
@@ -32,7 +32,7 @@ public void attachColorList(NSColorList colorList) {
 }
 
 public NSColor color() {
-	long /*int*/ result = OS.objc_msgSend(this.id, OS.sel_color);
+	long result = OS.objc_msgSend(this.id, OS.sel_color);
 	return result != 0 ? new NSColor(result) : null;
 }
 
@@ -41,15 +41,15 @@ public void setColor(NSColor color) {
 }
 
 public static NSColorPanel sharedColorPanel() {
-	long /*int*/ result = OS.objc_msgSend(OS.class_NSColorPanel, OS.sel_sharedColorPanel);
+	long result = OS.objc_msgSend(OS.class_NSColorPanel, OS.sel_sharedColorPanel);
 	return result != 0 ? new NSColorPanel(result) : null;
 }
 
-public static double /*float*/ minFrameWidthWithTitle(NSString aTitle, long /*int*/ aStyle) {
-	return (double /*float*/)OS.objc_msgSend_fpret(OS.class_NSColorPanel, OS.sel_minFrameWidthWithTitle_styleMask_, aTitle != null ? aTitle.id : 0, aStyle);
+public static double minFrameWidthWithTitle(NSString aTitle, long aStyle) {
+	return OS.objc_msgSend_fpret(OS.class_NSColorPanel, OS.sel_minFrameWidthWithTitle_styleMask_, aTitle != null ? aTitle.id : 0, aStyle);
 }
 
-public static long /*int*/ windowNumberAtPoint(NSPoint point, long /*int*/ windowNumber) {
+public static long windowNumberAtPoint(NSPoint point, long windowNumber) {
 	return OS.objc_msgSend(OS.class_NSColorPanel, OS.sel_windowNumberAtPoint_belowWindowWithWindowNumber_, point, windowNumber);
 }
